Allow collaborators to create, delete and edit visualizations.
