+++ redash/permissions.py
@@ -1,6 +1,7 @@
+import functools
+
 from flask_login import current_user
 from flask_restful import abort
-import functools
 from funcy import flatten
 
 view_only = True
