+++ tests/test_handlers.py
@@ -1,11 +1,10 @@
-from funcy import project
-
 from flask import url_for
 from flask_login import current_user
+from funcy import project
 from mock import patch
+from tests import BaseTestCase, authenticated_user
+
 from redash import models, settings
-from tests import BaseTestCase
-from tests import authenticated_user
 
 
 class AuthenticationTestMixin(object):
@@ -63,121 +62,6 @@ class StatusTest(BaseTestCase):
     def test_redirects_non_authenticated_user(self):
         rv = self.client.get('/status.json')
         self.assertEqual(rv.status_code, 302)
-
-
-class VisualizationResourceTest(BaseTestCase):
-    def test_create_visualization(self):
-        query = self.factory.create_query()
-        models.db.session.commit()
-        data = {
-            'query_id': query.id,
-            'name': 'Chart',
-            'description': '',
-            'options': {},
-            'type': 'CHART'
-        }
-
-        rv = self.make_request('post', '/api/visualizations', data=data)
-
-        self.assertEquals(rv.status_code, 200)
-        data.pop('query_id')
-        self.assertDictContainsSubset(data, rv.json)
-
-    def test_delete_visualization(self):
-        visualization = self.factory.create_visualization()
-        models.db.session.commit()
-        rv = self.make_request('delete', '/api/visualizations/{}'.format(visualization.id))
-
-        self.assertEquals(rv.status_code, 200)
-        self.assertEquals(models.db.session.query(models.Visualization).count(), 0)
-
-    def test_update_visualization(self):
-        visualization = self.factory.create_visualization()
-        models.db.session.commit()
-        rv = self.make_request('post', '/api/visualizations/{0}'.format(visualization.id), data={'name': 'After Update'})
-
-        self.assertEquals(rv.status_code, 200)
-        self.assertEquals(rv.json['name'], 'After Update')
-
-    def test_only_owner_or_admin_can_create_visualization(self):
-        query = self.factory.create_query()
-        other_user = self.factory.create_user()
-        admin = self.factory.create_admin()
-        admin_from_diff_org = self.factory.create_admin(org=self.factory.create_org())
-        models.db.session.commit()
-        models.db.session.refresh(admin)
-        models.db.session.refresh(other_user)
-        models.db.session.refresh(admin_from_diff_org)
-        data = {
-            'query_id': query.id,
-            'name': 'Chart',
-            'description': '',
-            'options': {},
-            'type': 'CHART'
-        }
-
-
-        rv = self.make_request('post', '/api/visualizations', data=data, user=admin)
-        self.assertEquals(rv.status_code, 200)
-
-        rv = self.make_request('post', '/api/visualizations', data=data, user=other_user)
-        self.assertEquals(rv.status_code, 403)
-
-        rv = self.make_request('post', '/api/visualizations', data=data, user=admin_from_diff_org)
-        self.assertEquals(rv.status_code, 404)
-
-    def test_only_owner_or_admin_can_edit_visualization(self):
-        vis = self.factory.create_visualization()
-        models.db.session.flush()
-        path = '/api/visualizations/{}'.format(vis.id)
-        data = {'name': 'After Update'}
-
-        other_user = self.factory.create_user()
-        admin = self.factory.create_admin()
-        admin_from_diff_org = self.factory.create_admin(org=self.factory.create_org())
-        models.db.session.commit()
-        models.db.session.refresh(admin)
-        models.db.session.refresh(other_user)
-        models.db.session.refresh(admin_from_diff_org)
-
-        rv = self.make_request('post', path, user=admin, data=data)
-        self.assertEquals(rv.status_code, 200)
-
-        rv = self.make_request('post', path, user=other_user, data=data)
-        self.assertEquals(rv.status_code, 403)
-
-        rv = self.make_request('post', path, user=admin_from_diff_org, data=data)
-        self.assertEquals(rv.status_code, 404)
-
-    def test_only_owner_or_admin_can_delete_visualization(self):
-        vis = self.factory.create_visualization()
-        models.db.session.flush()
-        path = '/api/visualizations/{}'.format(vis.id)
-
-        other_user = self.factory.create_user()
-        admin = self.factory.create_admin()
-        admin_from_diff_org = self.factory.create_admin(org=self.factory.create_org())
-
-        models.db.session.commit()
-        models.db.session.refresh(admin)
-        models.db.session.refresh(other_user)
-        models.db.session.refresh(admin_from_diff_org)
-        rv = self.make_request('delete', path, user=admin)
-        self.assertEquals(rv.status_code, 200)
-
-        vis = self.factory.create_visualization()
-        models.db.session.commit()
-        path = '/api/visualizations/{}'.format(vis.id)
-
-        rv = self.make_request('delete', path, user=other_user)
-        self.assertEquals(rv.status_code, 403)
-
-        vis = self.factory.create_visualization()
-        models.db.session.commit()
-        path = '/api/visualizations/{}'.format(vis.id)
-
-        rv = self.make_request('delete', path, user=admin_from_diff_org)
-        self.assertEquals(rv.status_code, 404)
 
 
 class JobAPITest(BaseTestCase, AuthenticationTestMixin):
